requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
dnspython==1.16.0
h11==0.16.0
idna==3.10
motor==3.7.1
pydantic_core==2.16.2
pymongo==4.13.2
sniffio==1.3.1
starlette==0.35.1
typing_extensions==4.14.1